style(ts): add typescript
