+++ src/pages/index.tsx
@@ -1,54 +1,54 @@
 import * as React from 'react';
 
 // styles
-const pageStyles = {
+const pageStyles: React.CSSProperties = {
   color: '#232129',
   padding: 96,
   fontFamily: '-apple-system, Roboto, sans-serif, serif',
 };
-const headingStyles = {
+const headingStyles: React.CSSProperties = {
   marginTop: 0,
   marginBottom: 64,
   maxWidth: 320,
 };
-const headingAccentStyles = {
+const headingAccentStyles: React.CSSProperties = {
   color: '#663399',
 };
-const paragraphStyles = {
+const paragraphStyles: React.CSSProperties = {
   marginBottom: 48,
 };
-const codeStyles = {
+const codeStyles: React.CSSProperties = {
   color: '#8A6534',
   padding: 4,
   backgroundColor: '#FFF4DB',
   fontSize: '1.25rem',
   borderRadius: 4,
 };
-const listStyles = {
+const listStyles: React.CSSProperties = {
   marginBottom: 96,
   paddingLeft: 0,
 };
-const listItemStyles = {
+const listItemStyles: React.CSSProperties = {
   fontWeight: 300,
   fontSize: 24,
   maxWidth: 560,
   marginBottom: 30,
 };
 
-const linkStyle = {
+const linkStyle: React.CSSProperties = {
   color: '#8954A8',
   fontWeight: 'bold',
   fontSize: 16,
   verticalAlign: '5%',
 };
 
-const docLinkStyle = {
+const docLinkStyle: React.CSSProperties = {
   ...linkStyle,
   listStyleType: 'none',
   marginBottom: 24,
 };
 
-const descriptionStyle = {
+const descriptionStyle: React.CSSProperties = {
   color: '#232129',
   fontSize: 14,
   marginTop: 10,
@@ -62,7 +62,7 @@ const docLink = {
   color: '#8954A8',
 };
 
-const badgeStyle = {
+const badgeStyle: React.CSSProperties = {
   color: '#fff',
   backgroundColor: '#088413',
   border: '1px solid #088413',
@@ -127,56 +127,58 @@ const links = [
 
 // markup
 function IndexPage() {
-  <main style={pageStyles}>
-    <title>Home Page</title>
-    <h1 style={headingStyles}>
-      Congratulations
-      <br />
-      <span style={headingAccentStyles}>— you just made a Gatsby site! </span>
-      <span role="img" aria-label="Party popper emojis">
-        🎉🎉🎉
-      </span>
-    </h1>
-    <p style={paragraphStyles}>
-      Edit <code style={codeStyles}>src/pages/index.js</code> to see this page
-      update in real-time.{' '}
-      <span role="img" aria-label="Sunglasses smiley emoji">
-        😎
-      </span>
-    </p>
-    <ul style={listStyles}>
-      <li style={docLinkStyle}>
-        <a
-          style={linkStyle}
-          href={`${docLink.url}?utm_source=starter&utm_medium=start-page&utm_campaign=minimal-starter`}
-        >
-          {docLink.text}
-        </a>
-      </li>
-      {links.map((link) => (
-        <li key={link.url} style={{ ...listItemStyles, color: link.color }}>
-          <span>
-            <a
-              style={linkStyle}
-              href={`${link.url}?utm_source=starter&utm_medium=start-page&utm_campaign=minimal-starter`}
-            >
-              {link.text}
-            </a>
-            {link.badge && (
-              <span style={badgeStyle} aria-label="New Badge">
-                NEW!
-              </span>
-            )}
-            <p style={descriptionStyle}>{link.description}</p>
-          </span>
+  return (
+    <main style={pageStyles}>
+      <title>Home Page</title>
+      <h1 style={headingStyles}>
+        Congratulations
+        <br />
+        <span style={headingAccentStyles}>— you just made a Gatsby site! </span>
+        <span role="img" aria-label="Party popper emojis">
+          🎉🎉🎉
+        </span>
+      </h1>
+      <p style={paragraphStyles}>
+        Edit <code style={codeStyles}>src/pages/index.js</code> to see this page
+        update in real-time.{' '}
+        <span role="img" aria-label="Sunglasses smiley emoji">
+          😎
+        </span>
+      </p>
+      <ul style={listStyles}>
+        <li style={docLinkStyle}>
+          <a
+            style={linkStyle}
+            href={`${docLink.url}?utm_source=starter&utm_medium=start-page&utm_campaign=minimal-starter`}
+          >
+            {docLink.text}
+          </a>
         </li>
-      ))}
-    </ul>
-    <img
-      alt="Gatsby G Logo"
-      src="data:image/svg+xml,%3Csvg width='24' height='24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M12 2a10 10 0 110 20 10 10 0 010-20zm0 2c-3.73 0-6.86 2.55-7.75 6L14 19.75c3.45-.89 6-4.02 6-7.75h-5.25v1.5h3.45a6.37 6.37 0 01-3.89 4.44L6.06 9.69C7 7.31 9.3 5.63 12 5.63c2.13 0 4 1.04 5.18 2.65l1.23-1.06A7.959 7.959 0 0012 4zm-8 8a8 8 0 008 8c.04 0 .09 0-8-8z' fill='%23639'/%3E%3C/svg%3E"
-    />
-  </main>;
+        {links.map((link) => (
+          <li key={link.url} style={{ ...listItemStyles, color: link.color }}>
+            <span>
+              <a
+                style={linkStyle}
+                href={`${link.url}?utm_source=starter&utm_medium=start-page&utm_campaign=minimal-starter`}
+              >
+                {link.text}
+              </a>
+              {link.badge && (
+                <span style={badgeStyle} aria-label="New Badge">
+                  NEW!
+                </span>
+              )}
+              <p style={descriptionStyle}>{link.description}</p>
+            </span>
+          </li>
+        ))}
+      </ul>
+      <img
+        alt="Gatsby G Logo"
+        src="data:image/svg+xml,%3Csvg width='24' height='24' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M12 2a10 10 0 110 20 10 10 0 010-20zm0 2c-3.73 0-6.86 2.55-7.75 6L14 19.75c3.45-.89 6-4.02 6-7.75h-5.25v1.5h3.45a6.37 6.37 0 01-3.89 4.44L6.06 9.69C7 7.31 9.3 5.63 12 5.63c2.13 0 4 1.04 5.18 2.65l1.23-1.06A7.959 7.959 0 0012 4zm-8 8a8 8 0 008 8c.04 0 .09 0-8-8z' fill='%23639'/%3E%3C/svg%3E"
+      />
+    </main>
+  );
 }
 
 export default IndexPage;
